Move metadata_store to the folder
