+++ src/tribler-core/tribler_core/components/gigachannel_manager.py
@@ -1,6 +1,6 @@
 from tribler_core.components.base import Component
 from tribler_core.components.libtorrent import LibtorrentComponent
-from tribler_core.components.metadata_store import MetadataStoreComponent
+from tribler_core.components.metadata_store.metadata_store_component import MetadataStoreComponent
 from tribler_core.components.reporter import ReporterComponent
 from tribler_core.components.restapi import RESTComponent
 from tribler_core.modules.metadata_store.manager.gigachannel_manager import GigaChannelManager
