+++ src/tribler-core/tribler_core/components/torrent_checker.py
@@ -1,7 +1,7 @@
 from tribler_common.simpledefs import STATE_START_TORRENT_CHECKER
 from tribler_core.components.base import Component
 from tribler_core.components.libtorrent import LibtorrentComponent
-from tribler_core.components.metadata_store import MetadataStoreComponent
+from tribler_core.components.metadata_store.metadata_store_component import MetadataStoreComponent
 from tribler_core.components.reporter import ReporterComponent
 from tribler_core.components.restapi import RESTComponent
 from tribler_core.components.socks_configurator import SocksServersComponent
